Simplified everything by adding a main and moving the creation of the tamagoshi to the Main and not GraphicTama
